requirements.txt
aiohttp==0.21.5
cffi==1.6.0
chardet==2.3.0
discord.py==0.10.0a0
pycparser==2.14
PyNaCl==1.0.1
requests==2.9.1
six==1.10.0
websockets==3.1
